Add support for RN59
Until .58 or .59, RN used JSObjectMakeFunctionWithCallback with the function pointer to create native-based JavaScript function wrappers. This changed in order to support the new JSI model. Instead, RN now creates a JS class which has three callbacks (initialize, finalize and callAsFunction). A function object is then created with this class, with an internal struct as the object's private data. This object is then associated as a property on the global JS object. When JS code calls this function, the callAsFunction callback is called; it extracts the private data from the object and calls the originally provided function implementation. These components are all disconnected from each other in a way that JSObjectMakeFunctionWithCallback wasn't. In the fix above, I found a way to connect these components back in a manner similar to how JSObjectMakeFunctionWithCallback works, so that the same logic as before can be used.
#40